Enable windows teaserpp build

+++ 3rdparty/teaser_plusplus/teaser/include/teaser/linalg.h
@@ -63,8 +63,8 @@ Eigen::Matrix<NumT, Eigen::Dynamic, 1> vectorKron(const Eigen::Matrix<NumT, N, 1
                                                   const Eigen::Matrix<NumT, M, 1>& v2) {
   Eigen::Matrix<double, Eigen::Dynamic, 1> output(v1.rows() * v2.rows(), 1);
 #pragma omp parallel for collapse(2) shared(v1, v2, output) default(none)
-  for (size_t i = 0; i < v1.rows(); ++i) {
-    for (size_t j = 0; j < v2.rows(); ++j) {
+  for (int i = 0; i < v1.rows(); ++i) {
+    for (int j = 0; j < v2.rows(); ++j) {
       output[i * v2.rows() + j] = v1[i] * v2[j];
     }
   }
